requirements.txt
discord.py
yt-dlp
youtube-search-python
pytube
tqdm
colorama
python-dotenv
